This is the first fully working version of ffccnx (CCNx enabled Firefox browswer).
  An critical bug was fixed in nsCCNxInputStream now that any size of file in repo
can be transmitted to the browser. Now html, video, image are all supported just like
the `file` protocol.

  Without an equivalent to HTTP protocol, the only way to support MIME type infrerence
is the use nsIMIMEService::GetTypeFromURI. This method is purely based on the file
name extention, which is not really reliable and should be modified in future. A
reasonable solution might be include content type information (even codec information
for video and audio), in the interest sending back to the user side by the server.
But anyway, a server is definitely need.